[add]一致度の計算を外部に切り出し
