+++ src/hooks/createNetwork.ts
@@ -4,10 +4,11 @@ import { DEFAULT_FILTER } from '@/constants/DEFAULT_FILTER';
 import { Filter } from '@/types/api/FilterType';
 import { SteamDetailsDataType, SteamDeviceType, SteamGenreType } from '@/types/api/getSteamDetailType';
 import { ISR_FETCH_INTERVAL } from '@/constants/DetailsConstants';
+import { calcAllMatchPercentage } from '@/components/common/CalcMatch';
 
 type Node = {
   circleScale?: number,
-  gameData: GameData,
+  gameData: SteamDetailsDataType,
   id: number,
   imgURL: string,
   index?: number,
@@ -21,66 +22,6 @@ type Node = {
   y?: number,
 }
 
-type GameData = {
-  genres : SteamGenreType[],
-  price: number,
-  isSinglePlayer: boolean,
-  isMultiPlayer: boolean,
-  device: SteamDeviceType,
-}
-
-const calcGenresPercentage = (filter: Filter, genres: SteamGenreType[]) => {
-  let genreCount = 0;
-
-  genres.forEach((genre: SteamGenreType) => {
-    if(filter.Categories[genre.id]) {
-      genreCount++;
-    }
-  });
-
-  return (genreCount / genres.length) * 100;
-}
-
-const calcPricePercentage = (filter: Filter, price: number) => {
-  const startPrice = filter.Price.startPrice;
-  const endPrice = filter.Price.endPrice;
-  const diffPrice = endPrice - startPrice;
-  const diffScale = d3.scaleLinear()
-                  .domain([0, diffPrice])
-                  .range([100, 0]);
-
-  if(startPrice <= price && price <= endPrice) {
-    return 100;
-  } else if(price < startPrice) {
-    const diff = startPrice - price;
-    return diffScale(diff) >= 0 ? diffScale(diff) : 0;
-  } else {
-    const diff = price - endPrice;
-    return diffScale(diff) >= 0 ? diffScale(diff) : 0;
-  }
-}
-
-const calcModePercentage = (filter: Filter, modes: {isSinglePlayer: boolean, isMultiPlayer: boolean}) => {
-  let modeCount = 0;
-  Object.keys(modes).forEach((mode: string) => {
-    if(mode === "isSinglePlayer" && filter.Mode.isSinglePlayer || mode === "isMultiPlayer" && filter.Mode.isMultiPlayer) {
-      modeCount++;
-    }
-  });
-
-  return (modeCount / 2) * 100;
-}
-
-const calcDevicePercentage = (filter: Filter, devices: SteamDeviceType) => {
-  let deviceCount = 0;
-  Object.keys(devices).forEach((device: string) => {
-    if((device === "windows" && filter.Device.windows) || (device === "mac" && filter.Device.mac)) {
-      deviceCount++;
-    }
-  });
-
-  return (deviceCount / 2) * 100;
-}
 
 const calcCommonGenres = (game1:any, game2:any) => {
   let genresWeight = 1;
@@ -221,21 +162,8 @@ const createNetwork = async () => {
   });
 
   nodes.forEach((node: Node) => {
-    // 一致度を計算（ジャンル）
-    const genreMatchPercent = calcGenresPercentage(filter, node.gameData.genres);
-
-    // 一致度を計算(価格)
-    const priceMatchPercent = calcPricePercentage(filter, node.gameData.price);
-
-    // 一致度を計算(ゲームモード)
-    const modeMatchPercent = calcModePercentage(filter, {isSinglePlayer: node.gameData.isSinglePlayer, isMultiPlayer: node.gameData.isMultiPlayer});
-
-    // 一致度を計算(対応デバイス)
-    const deviceMatchPercent = calcDevicePercentage(filter, node.gameData.device);
-
     // 一致度を計算(全体)
-    const overallMatchPercent = parseFloat(((genreMatchPercent + priceMatchPercent + modeMatchPercent + deviceMatchPercent) / 4).toFixed(1));
-
+    const overallMatchPercent = calcAllMatchPercentage(filter, node.gameData);
     node.circleScale = matchScale(overallMatchPercent);
   });
